requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
cymem==2.0.8
dj-database-url==2.3.0
Django==5.1.3
docx2txt==0.8
fastjsonschema==2.19.1
Flask==3.1.0
flatbuffers==24.3.25
gast==0.6.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
json5==0.9.14
keras==3.6.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.7
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.7
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodome==3.19.1
Pygments==2.17.2
pyparsing==3.1.1
pyperclip==1.9.0
pyproject_hooks==1.0.0
pyrsistent==0.20.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.16.2
rsa==4.9
scapy==2.5.0
Send2Trash==1.8.2
setuptools==69.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
stats==0.1.2a0
sympy==1.12
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wheel==0.42.0
widgetsnbextension==4.0.9
wrapt==1.16.0
zipp==3.17.0
